fix: null totalPartnerCost, Scheduled list filter, webhook bounded read
Follow-up fixes from the code review of the Phase 5 (int->long) work:

- CampaignCounters.TotalPartnerCost is now decimal? — the API returns
  null for it and StringToDecimalConverter threw on the null token,
  crashing Campaign.GetInfoAsync.
- ApiStatusCodes.ToApiCode now maps SmsStatus.Scheduled to "SCHEDUL";
  Message.ListAsync(Status = Scheduled) previously threw before the HTTP
  request.
- The webhook endpoint now enforces MaxRequestBodyBytes via a bounded
  read, so a chunked or absent Content-Length can no longer bypass the cap.
- Docs: corrected ApiUrl examples (must not include /service/, which the
  client appends) and README campaign/task id types (int -> long).
- Consistency: widened ContactCardFilterSpec.GroupId and
  AddressFilterSpec.RegionId/CityId to long? to match the long entity ids
  (not a correctness bug — the LINQ filter already handled 64-bit ids).

All 226 tests pass; build clean (0 warnings).

+++ src/Mobizon.Contracts/Models/ContactCards/AddressFilterSpec.cs
@@ -16,13 +16,13 @@
         public string Country { get; } = string.Empty;
 
         /// <summary>Region / province ID.</summary>
-        public int? RegionId { get; }
+        public long? RegionId { get; }
 
         /// <summary>Region / province name.</summary>
         public string Region { get; } = string.Empty;
 
         /// <summary>City ID.</summary>
-        public int? CityId { get; }
+        public long? CityId { get; }
 
         /// <summary>City name.</summary>
         public string City { get; } = string.Empty;
